Added exception handling with tests: get/delete non-existent game. Required switch from legacy Query.get() method to Session.get().

+++ app/crud/games.py
@@ -8,8 +8,8 @@ class NotFoundError(Exception):
     pass
 
 
-def get_game(session: Session, game_id: int, ) -> Game:
-    game = session.query(DBGame).get(game_id)
+def get_game(session: Session, game_id: int) -> Game:
+    game = session.get(DBGame, game_id)
     if game is None:
         raise NotFoundError
     return game
@@ -29,7 +29,9 @@ def create_game(session: Session, params: GameCreate) -> Game:
 
 
 def update_game(session: Session, game_id: int, params: GameUpdate) -> Game:
-    game = session.query(DBGame).get(game_id)
+    game = session.get(DBGame, game_id)
+    if game is None:
+        raise NotFoundError
     for attr, value in params.model_dump().items():
         setattr(game, attr, value)
     session.add(game)
@@ -39,7 +41,9 @@ def update_game(session: Session, game_id: int, params: GameUpdate) -> Game:
 
 
 def delete_game(session: Session, game_id: int) -> Game:
-    game = session.query(DBGame).get(game_id)
+    game = session.get(DBGame, game_id)
+    if game is None:
+        raise NotFoundError
     session.delete(game)
     session.commit()
     return game
